refactors EvolutionChain components
